Updates for UKB data
I've added in the UKB data. Currently, the app only works for genes that
have PTVs in gnomAD (so I'm not including ~200 genes that don't have
PTVs in gnomAD but do in UKB). If a gene doesn't have PTVs in UKB, it
has zero for the allele frequency and CI for now.